simulation/pyproject.toml
[project]
name = "tarware-simulation"
version = "0.1.0"
description = "TA-RWARE simulation server with headless rendering and SSE frame streaming"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "pillow==10.3.0",
    "numpy==1.26.4",
    "requests==2.32.3",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",   # for testing FastAPI endpoints
    "ruff",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
